+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change to default `EnforcedStyle: link_or_button` for `Capybara/ClickLinkOrButtonStyle` cop. ([@ydah])
 - Fix a false negative for `RSpec/HaveSelector` when first argument is dstr node. ([@ydah])
+- Fix an invalid attributes parse when name with multiple `[]` for `Capybara/SpecificFinders` and `Capybara/SpecificActions` and `Capybara/SpecificMatcher`. ([@ydah])
 
 ## 2.19.0 (2023-09-20)